Create FinallyBlock.java
The finally block in java is used to put important codes such as closing the connection, etc. The finally block executes whether exception rise or not and whether exception handled or not .  finally is useful for more than just exception handling — it allows the programmer to avoid having cleanup code accidentally bypassed by a return , continue , or break . Putting cleanup code in a finally block is always a good practice, even when no exceptions are anticipated.